offers-microservice/requirements.txt
django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-prometheus
black==24.3.0
flake8==7.0.0
whitenoise==6.6.0